+++ cli/kpSift3D.c
@@ -10,7 +10,7 @@
 
 #include <stdio.h>
 #include <getopt.h>
-#include "macros.h"
+#include "immacros.h"
 #include "imutil.h"
 #include "sift.h"